Error handling and expanded the responsibilities of the address bar
The address bar was already designed to be a self-contained entity that's responsible for navigating directories and reporting if they're valid. It was designed this way to unload the responsibility of manually navigating and verifying paths from other code. This expands on its responsibilities to report more information about a directory and the files contained in it, as well as adds proper error handling on directory and file access.