Change how rules are handled and processed
Rules no longer track whether they modify a file or not based on selection. Instead, each file can individually be enabled or disabled for renaming. Added iconography to support this and other modes a file can be in. Also did a cleanup of how rules and files are processed in general, to make the program-flow easier to understand. The beginnings of reordering rules by dragging has also been implemented.